third_party/WebKit/LayoutTests/inspector/tracing/tracing-timeline-load-expected.txt

Tests tracing based Timeline save/load functionality.


Running: testNormal

Running: testBroken

Running: testMalformedJSON
